arch: arm: aarch32: Split the basic stack frame
The Cortex-R return state should be the first thing saved on the stack
which means that it should be at the bottom of the esf struct.  However,
when the FPU is enabled the return state does not line up with its
location in the esf struct.  Split the basic stack frame so that on
Cortex-R the return state can be at the end.